+++ migrations/000022_secret_type.up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE secret_values ADD COLUMN type text NOT NULL DEFAULT 'string';